to me very important program
This is time saving, very important. Once you have typed a German = English vocabulary you can drop it in this program to flip it around. For example, you can insert (open) the 'de-eng-long.txt' file and run the program, which will produce a flipped vocabulary (dictionary): the English = German version.