working version
This is the secondary program which is still on the drawing board. To be replaced later with a newer version. Not significant changes are anticipated. It can be used to test foreign language vocabularies that user has already created in the form (in German-English), e.g. Buch, s (pl. Bücher) = Book